pyproject.toml
[tool.poetry]
name = "lang-chain-test"
version = "0.1.0"
description = ""
authors = ["naremloa <naremloa@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
llama-index = "^0.5.23"
flask = "^2.2.3"
python-dotenv = "^1.0.0"
bs4 = "^0.0.1"
fake-useragent = "^1.1.3"
selenium = "^4.9.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
